Fixed a bug after registration
I had a method in my HomePage class which is called loadUserData and this method was helping me for fetch the data from firebase to UserProfile. Today i realized that if the user log out nad register again a new account the method was calling prevoius user's data's until restrart the app.